moved, community and profile pages
