+++ frontend/src/app/features/hubit/community/community.component.html
@@ -1,3 +1,15 @@
+<div *ngIf="show_follow_button">
+  <div *ngIf="user_following; else followCommunity">
+    <button mat-button (click)="unfollowCommunityCall()" class="follow-toggle">
+      Unfollow
+    </button>
+  </div>
+</div>
+<ng-template #followCommunity>
+  <button mat-button (click)="followCommunityCall()" class="follow-toggle">
+    Follow
+  </button>
+</ng-template>
 <a href="/new_post?community={{ community_id }}" class="new-post-button"
   ><button mat-button>New Post</button></a
 >
@@ -6,28 +18,39 @@
     <p class="community-title">{{ community_name }}</p>
     <p class="community-description">{{ community_desc }}</p>
   </div>
+  <div class="boards-button">
+    <a
+      [routerLink]="['/community', community_id, 'boards']"
+      class="boards-button"
+    >
+      <button mat-button>Boards</button>
+    </a>
+  </div>
   <div class="post_list">
     <li *ngFor="let post of community_posts; index as i">
-      <mat-card class="post_card">
-        <mat-card-header>
-          <mat-card-title>
-            {{ post.created_by_username }}
-          </mat-card-title>
-          <mat-card-subtitle>
-            {{ community_name }}
-          </mat-card-subtitle>
-        </mat-card-header>
-        <mat-card-content>
-          <p>
-            {{ post.content }}
-          </p>
-        </mat-card-content>
-        <mat-card-actions>
-          <button mat-button>LIKE</button>
-          <button mat-button>DISLIKE</button>
-          <button mat-button>SHARE</button>
-        </mat-card-actions>
-      </mat-card>
+      <a href="/post?post={{ post.id }}">
+        <mat-card class="post_card">
+          <mat-card-header>
+            <mat-card-title>
+              <a href="/profile?id={{ post.created_by.user.id }}">{{
+                post.created_by ? post.created_by.community_username : "_null"
+              }}</a>
+            </mat-card-title>
+            <mat-card-subtitle>
+              {{ community_name }}
+            </mat-card-subtitle>
+          </mat-card-header>
+          <mat-card-content>
+            <p>
+              {{ post.content }}
+            </p>
+          </mat-card-content>
+          <mat-card-actions>
+            <button mat-button>LIKE</button>
+            <button mat-button>SHARE</button>
+          </mat-card-actions>
+        </mat-card>
+      </a>
     </li>
   </div>
 </div>
