+++ frontend/src/app/features/hubit/followed-communities/followed-communities.component.ts
