+++ frontend/src/app/features/hubit/new_post/new_post.component.css
@@ -1,3 +1,7 @@
+* {
+  /*border: 1px solid black;*/
+}
+
 .middle_box {
   position: absolute;
   left: 10vw;
@@ -22,7 +26,7 @@
 
 .middle_box .post_form .input-80-width .post-text {
   width: 100%;
-  border: 2px solid rgb(30, 30, 30);
+  border: 2px solid rgb(30,30,30);
   border-radius: 15px;
   height: 20vh;
 
@@ -32,20 +36,6 @@
   font-family: Verdana, Geneva, Tahoma, sans-serif;
   overflow-y: auto;
   resize: none;
-}
-
-.middle_box .post_form .input-80-width .image-preview-box {
-  width: 100%;
-  height: 10vh;
-  line-height: 10vh;
-  background-color: rgb(217, 217, 217);
-  text-align: center;
-}
-
-.image-preview-box .image-preview-box-text {
-  font-size: 1em;
-  display: inline-block;
-  vertical-align: middle;
 }
 
 .middle_box .post-button-area {
@@ -58,7 +48,7 @@
   width: 15%;
   height: 60%;
   font-size: 2em;
-  background-color: rgb(210, 210, 210);
+  background-color: rgb(210,210,210);
 }
 
 .spacer {
