+++ frontend/src/app/features/hubit/profile/profile.component.html
@@ -1,59 +1,71 @@
-<div class="profile-container">
-  <div class="info-container">
-    <div class="profile-picture">
-      <mat-icon class="person_icon">account_circle</mat-icon>
-    </div>
-
-    <div class="details">
+<div *ngIf="logged_in && !viewing_own_profile">
+  <a href="/" class="direct-message-button"><button mat-button>Direct Message</button></a>
+</div>
+<div class="center-flex">
+<div class="profile_grid">
+  <div class="col">
+    <mat-icon class="person_icon">account_circle</mat-icon>
+  </div>
+  <div class="col">
       <h1 class="username">
-        {{ username }}
+          {{username}}
       </h1>
-      <p class="bio">{{ bio }}</p>
-      <div class="stats-container">
-        <div class="stat">
-          <h3 class="profile_number">{{ num_posts }}</h3>
-          <p class="profile_number_text">Posts</p>
-        </div>
-        <div class="stat">
-          <h3 class="profile_number">{{ num_followers }}</h3>
-          <p class="profile_number_text">Followers</p>
-        </div>
-        <div class="stat">
-          <h3 class="profile_number">{{ num_following }}</h3>
-          <p class="profile_number_text">Following</p>
-        </div>
-        <div class="stat">
-          <h3 class="profile_number">{{ num_communities }}</h3>
-          <p class="profile_number_text">Communities</p>
-        </div>
-      </div>
-    </div>
+      <p class="bio">{{bio}}</p>
   </div>
 
-  <div class="divider"></div>
-
-  <div class="post_list">
-    <!-- Use material cards: https://material.angular.io/components/card/overview -->
-    <li *ngFor="let post of posts; index as i">
-      <mat-card class="post_card">
-        <mat-card-header>
-          <mat-card-title>
-            {{ username }}
-          </mat-card-title>
-          <mat-card-subtitle>
-            <!-- Fix this later -->
-            community_not_set
-          </mat-card-subtitle>
-        </mat-card-header>
-        <mat-card-content>
-          <p>Post {{ i }}. {{ post.content }}</p>
-        </mat-card-content>
-        <mat-card-actions>
-          <button mat-button>LIKE</button>
-          <button mat-button>DISLIKE</button>
-          <button mat-button>SHARE</button>
-        </mat-card-actions>
-      </mat-card>
-    </li>
+  <div class="col">
+    <h3 class="profile_number">{{ num_posts }}</h3>
+    <p class="profile_number_text">Posts</p>
   </div>
+  <div class="col">
+    <h3 class="profile_number">{{ num_followers }}</h3>
+    <p class="profile_number_text">Followers</p>
+  </div>
+  <div class="col">
+    <h3 class="profile_number">{{ num_following }}</h3>
+    <p class="profile_number_text">Following</p>
+  </div>
+  <div class="col" (click)="toFollowedCommunities()">
+    <h3 class="profile_number">{{ followed_communities.length }}</h3>
+    <p class="profile_number_text">Communities</p>
+  </div>
+</div>
+
+<!-- If viewing own profile -->
+<div>
+  <div class="col" *ngIf="viewing_own_profile">
+    <button
+      mat-raised-button
+      class="settings_button"
+      (click)="settings_action()"
+    >
+      <mat-icon class="settings_icon">settings</mat-icon> Settings
+    </button>
+  </div>
+</div>
+</div>
+<div class="post_list">
+<!-- Use material cards: https://material.angular.io/components/card/overview -->
+<li *ngFor="let post of posts; index as i">
+  <a href="post?post={{ post.id }}">
+    <mat-card class="post_card">
+      <mat-card-header>
+        <mat-card-title>
+          {{ username }}
+        </mat-card-title>
+        <mat-card-subtitle>
+          <a
+            href="community?community={{ post.created_by.for_community.id }}"
+            >{{ post.created_by.for_community.name }}</a
+          >
+        </mat-card-subtitle>
+      </mat-card-header>
+      <mat-card-content>
+        <p>
+          {{ post.content }}
+        </p>
+      </mat-card-content>
+    </mat-card>
+  </a>
+</li>
 </div>
