--- frontend/src/app/new-community/new-community.component.html
@@ -1,52 +1,0 @@
-<div class="middle_box">
-    <form class="new_community_form">
-        <h1 class="title-text">Create a New Community</h1>
-        <mat-form-field class="input-80-width">
-            <mat-label>Community Name:</mat-label>
-                    <input #community_name matInput>
-        </mat-form-field>
-
-        <mat-form-field class="input-80-width">
-            <h1>Description:</h1>
-            <textarea matInput #community_desc class="description-text"></textarea>
-        </mat-form-field>
-        
-        <mat-form-field class="input-80-width">
-            <mat-label>Choose Moderators</mat-label>
-
-            <mat-chip-grid [disabled]="hide_chips" #chipGrid>
-                <mat-chip-row *ngFor="let mod of selected_mods" (removed)="remove_mod(mod)">
-                    {{mod.username}}
-                    <button matChipRemove *ngIf="mod.id != this.self_id">
-                        <mat-icon>cancel</mat-icon>
-                    </button>
-                </mat-chip-row>
-            </mat-chip-grid>
-            
-            <input placeholder="New Mod..." #modInput [formControl]="userCtrl"
-            [matChipInputFor]="chipGrid" [matAutocomplete]="auto"
-            (input)="filter_users(modInput.value)"
-            [matChipInputSeparatorKeyCodes]="separatorKeyCodes"
-            (matChipInputTokenEnd)="add_mod($event)"/>
-
-            <mat-autocomplete #auto="matAutocomplete" (optionSelected)="select_mod($event)">
-                <mat-option *ngFor="let mod of filtered_mods | async" [value]="mod">
-                    {{mod}}
-                </mat-option>
-            </mat-autocomplete>
-        </mat-form-field>
-
-        <mat-form-field class="input-80-width">
-            <input matInput hidden="true">
-            <h1>Add Background Photo:</h1>
-            <!--<input matInput #Image type="file" (change)="onChange($event.target.files)">-->
-            <input #Image type="file" (change)="onChange($event)">
-        </mat-form-field>
-    
-        <div class="create-community-button-area">
-                                                            <!-- Couldn't figure out chip values, will fix later -->
-            <button mat-raised-button class="create-community-button" (click)="create_community(community_name.value, community_desc.value, ['General'])">Create Community</button>
-        </div>
-    </form>
-    <div class="spacer"></div>
-</div>
