--- frontend/src/app/new_post/new_post.component.css
@@ -1,58 +1,0 @@
-* {
-    /*border: 1px solid black;*/
-}
-
-.middle_box {
-    position: absolute;
-    left: 10vw;
-    width: 80vw;
-    top: 15vh;
-    height: auto;
-    margin-bottom: 10px;
-
-    border-radius: 50px;
-    background-color: white;
-    box-shadow: 0px 0px 10px gray;
-}
-
-.middle_box .post_form {
-    height: 100%;
-}
-
-.middle_box .post_form .input-80-width {
-    margin-left: 10%;
-    width: 80%;
-}
-
-.middle_box .post_form .input-80-width .post-text {
-    width: 100%;
-    border: 2px solid rgb(30,30,30);
-    border-radius: 15px;
-    height: 20vh;
-
-    padding: 5px;
-    line-height: 1.4;
-    font-size: 1.2rem;
-    font-family: Verdana, Geneva, Tahoma, sans-serif;
-    overflow-y: auto;
-    resize: none;
-}
-
-.middle_box .post-button-area {
-    width: 100%;
-    text-align: center;
-    height: 10vh;
-}
-
-.middle_box .post-button-area .post-button {
-    width: 15%;
-    height: 60%;
-    font-size: 2em;
-    background-color: rgb(210,210,210);
-}
-
-.spacer {
-    height: 5vh;
-    margin: 0 0 -5vh 0;
-    background: transparent;
-}
