--- frontend/src/app/profile/profile.component.css
@@ -1,190 +1,0 @@
-.direct-message-button {
-  position: absolute;
-  top: 14px;
-  right: 30px;
-  background-color: #ffffff;
-  border-radius: 5px;
-  color: #000000;
-}
-
-/* https://css-tricks.com/snippets/css/complete-guide-grid/ */
-
-.profile_grid {
-  display: grid;
-  grid-template-columns: repeat(12, 1fr);
-  grid-template-rows: [row1-start] 50vh [row1-end row2-start] 10vh [row2-end];
-  justify-items: start;
-
-  /* Can also use grid-template-area for defined areas of grid */
-
-  /* Gaps between cells: column-gap:10px; row-gap:15px; */
-  /* Will probably want justify-items: stretch; so items fill whole row, other options are: start, end, center */
-}
-
-/* For Icon/Image, Name, and Description */
-.col:nth-child(1) {
-  grid-column: 1 / span 6;
-  width: 50%;
-  height: 100%;
-  justify-self: end;
-  align-self: center;
-}
-.person_icon {
-  scale: 10;
-  position: relative;
-  top: 50%;
-  left: 50%;
-}
-
-.col:nth-child(2) {
-  grid-column: 7 / span 6;
-  /* Options are start, end, center, stretch */
-  justify-self: start; /* horizontal */
-  align-self: center; /* vertical */
-  width: 50%;
-  height: 60%;
-}
-.username {
-  font-size: xx-large;
-  font-weight: bolder;
-}
-.username_edit {
-  background-color: rgba(0, 0, 0, 0);
-  width: 100%;
-
-  font-size: 2em;
-  font-weight: bolder;
-  height: fit-content;
-
-  outline: none;
-  border: 1px solid gray;
-  /*border: none;*/
-
-  padding: 0px;
-  margin: 0px;
-}
-.bio {
-  line-height: 24px;
-  height: 75%;
-  overflow-y: auto;
-}
-.bio_edit {
-  background-color: rgba(0, 0, 0, 0);
-  border: 1px solid gray;
-  width: 100%;
-  outline: none;
-  /*border: none;*/
-  height: 75%;
-  overflow-y: auto;
-
-  padding: 0px;
-  margin: 0px;
-
-  font-family: Verdana, Geneva, Tahoma, sans-serif;
-  resize: none;
-}
-.col .cancel_button {
-  margin-left: 2.5vw;
-  background-color: crimson;
-  color: white;
-}
-.col .confirm_button {
-  margin-left: 2.5vw;
-  background-color: #17b169;
-  color: white;
-}
-
-/* For # posts, followers, following, and communities */
-.col:nth-child(3) {
-  grid-column: 5 / span 1;
-  justify-self: center;
-}
-.col:nth-child(4) {
-  grid-column: 6 / span 1;
-  justify-self: center;
-}
-.col:nth-child(5) {
-  grid-column: 7 / span 1;
-  justify-self: center;
-}
-.col:nth-child(6) {
-  grid-column: 8 / span 1;
-  justify-self: center;
-}
-.profile_number {
-  font-size: xx-large;
-  font-weight: 600;
-  font-family: Verdana, Geneva, Tahoma, sans-serif;
-  text-align: center;
-}
-.profile_number_text {
-  text-align: center;
-  font-size: large;
-}
-
-/* For viewing own profile buttons */
-.col:nth-child(7) {
-  grid-row: 3 / span 1;
-  grid-column: 5 / span 2;
-  justify-self: center;
-}
-.edit_button {
-  border-radius: 20px;
-  width: 110%;
-}
-.col:nth-child(8) {
-  grid-column: 7 / span 1;
-  justify-self: center;
-}
-.share_button {
-  border-radius: 20px;
-}
-.share_button .share_icon {
-  margin-left: 0.3vw;
-}
-.col:nth-child(9) {
-  grid-column: 8 / span 1;
-}
-.settings_button {
-  border-radius: 20px;
-}
-.settings_button .settings_icon {
-  margin-left: 0.3vw;
-}
-
-.col {
-  font-family: Verdana, Geneva, Tahoma, sans-serif;
-  width: 100%;
-  justify-items: center;
-}
-
-a {
-  text-decoration: none;
-  color: black;
-}
-
-.post_list {
-  /* https://css-tricks.com/snippets/css/a-guide-to-flexbox/ */
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100vh;
-  row-gap: 15px;
-  list-style-type: none;
-}
-
-.post_list .post_card {
-  width: 50vw;
-  align-self: center;
-}
-
-/* .create_new_community {
-  margin-left: 50px; 
-} */
-
-.center-flex {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 30px;
-}
